pyproject.toml
[tool.poetry]
name = "martel_printer_test_lib"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "martel_printer_test_lib"}]

[tool.poetry.dependencies]
python = "^3.10"
logic2_automation = { url = "https://saleae.github.io/logic2-automation/_downloads/3a6389974664c88edfbd009a9eaac439/logic2_automation-1.0.0-py3-none-any.whl" }
pywin32 = "^304"
Pillow = "^9.2.0"
pyserial = "^3.5"
pytest = "^7.1.3"
robotframework = "^5.0.1"
hanging-threads = "^2.0.5"

[tool.poetry.dev-dependencies]
pylint = "^2.15.3"
autopep8 = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

